LPS-123065 Moves the Calculator component to the editor folder
